Patch 0.1.9, solving instabilities
This is an emergency patch. Unfortunately, the API version previously used by the package started to be disabled by the Ipea servers in Brasilia and will eventually be completely discontinued. To keep the package working, I was forced to use the most up-to-date version of the API (http://www.ipeadata.gov.br/api/odata4/). Paradoxically, this newer version returns much less information than the old version. Because of this, some functionality previously existing in some functions of the package had to be sacrificed: i) the new version of the API does not allow the existence of the sources() function. The function has been kept in the package using the old version of the API, but it is subject to instabilities and can be completely broken at any time; ii) the themes() function no longer returns the "MACRO", "REGIONAL" and "SOCIAL" indicator columns; iii) the timeseries() function no longer has the "groupby" option.